extract min cut code to own module
